# Use collapse::pivot in resample.R functions for better performance