--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -62,7 +62,8 @@ Suggests:
     rmarkdown,
     testthat (>= 3.0.0),
     patchwork,
-    progress
+    progress,
+    collapse
 Config/testthat/edition: 3
 URL: https://github.com/tractometry/tractable
 BugReports: https://github.com/tractometry/tractable/issues