I noticed the functions in `resample.R` could probably be sped up by using functions from the package [`collapse`](https://cran.r-project.org/package=collapse). We will need a little more testing but here is a draft. 

Benchmark and results show several improvements in throughput and memory usage:

```
> devtools::load_all()

> mk_df <- function(n_subjectID=20000L, n_nodeID=40L, n_group=2L,seed=0) {
+ 	if(!is.null(seed)) {
+ 		set.seed(seed)
+ 	}
+ 	expand.grid(subjectID=se .... [TRUNCATED] 

> df1 <- mk_df(n_subjectID=1000L, n_nodeID = 10L, n_group=2L)

> df1_factor <- df1 |> dplyr::mutate(nodeID=as.factor(nodeID))

> df2 <- mk_df(n_subjectID=20000L, n_nodeID = 40L, n_group=2L)

> df2_factor <- df2 |> dplyr::mutate(nodeID=as.factor(nodeID))

> # Compare times
> # equality guaranteed by setting tidyr::pivot_longer/tidyr::pivot_wider row order options
> t1 <- bench::mark(
+ 	dplyr = withr::w .... [TRUNCATED] 

> t1
# A tibble: 2 × 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  
1 dplyr        5.71ms   5.87ms      169.    5.54MB     6.43    79     3      467ms <tibble>
2 collapse    787.4µs 858.95µs     1121.    2.87MB    15.4    510     7      455ms <tibble>
# ℹ 3 more variables: memory <list>, time <list>, gc <list>

> # Compare times, don't check for equality (result order may differ despite same seed)
> # Also compare against using a factor nodeID, which should b .... [TRUNCATED] 

> t1_nocheck
# A tibble: 3 × 13
  expression           min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result
  <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>
1 dplyr              5.8ms   5.89ms      169.    4.03MB     8.88    76     4      450ms <NULL>
2 collapse         776.5µs 853.02µs     1170.    1.38MB    17.6    532     8      455ms <NULL>
3 collapse_factor  662.2µs 727.26µs     1373.    1.34MB    20.3    609     9      443ms <NULL>
# ℹ 3 more variables: memory <list>, time <list>, gc <list>

> # Compare times
> # equality guaranteed by setting tidyr::pivot_longer/tidyr::pivot_wider row order options
> t2 <- bench::mark(
+ 	dplyr = withr::w .... [TRUNCATED] 

> t2
# A tibble: 2 × 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  
1 dplyr         270ms  271.6ms      3.68     287MB     14.7     2     8      543ms <tibble>
2 collapse     27.9ms   33.2ms     19.7      105MB     17.7    10     9      509ms <tibble>
# ℹ 3 more variables: memory <list>, time <list>, gc <list>

> rm(list=c("df1", "df1_factor"))

> # Compare times, don't check for equality (result order may differ despite same seed)
> # Also compare against using a factor nodeID, which should b .... [TRUNCATED] 

> t2_nocheck
# A tibble: 3 × 13
  expression           min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result
  <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>
1 dplyr            188.7ms  196.7ms      4.58   293.3MB     12.2     3     8      655ms <NULL>
2 collapse          28.6ms     33ms     24.9    105.3MB     24.9    13    13      523ms <NULL>
3 collapse_factor   22.8ms   28.1ms     29.2     99.2MB     21.4    15    11      513ms <NULL>
# ℹ 3 more variables: memory <list>, time <list>, gc <list>

> rm(list=c("df2", "df2_factor"))

> benchmarkme::get_cpu()
$vendor_id
character(0)

$model_name
[1] "Apple M1"

$no_of_cores
[1] 8


> benchmarkme::get_ram()
17.2 GB

> sessionInfo()
R version 4.4.2 (2024-10-31)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.7.2

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Los_Angeles
tzcode source: internal

attached base packages:
[1] stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
[1] tractable_0.2.1 testthat_3.2.3  felp_0.6.0     

loaded via a namespace (and not attached):
 [1] tidyselect_1.2.1      dplyr_1.1.4           farver_2.1.2          fastmap_1.2.0        
 [5] promises_1.3.2        digest_0.6.37         mime_0.12             lifecycle_1.0.4      
 [9] ellipsis_0.3.2        survival_3.8-3        magrittr_2.0.3        compiler_4.4.2       
[13] rlang_1.1.4           tools_4.4.2           utf8_1.2.4            collapse_2.0.19      
[17] fitdistrplus_1.2-2    htmlwidgets_1.6.4     pkgbuild_1.4.6        curl_6.1.0           
[21] bench_1.1.4           aws.signature_0.6.0   xml2_1.3.6            RColorBrewer_1.1-3   
[25] pkgload_1.4.0         miniUI_0.1.1.1        withr_3.0.2           purrr_1.0.2          
[29] desc_1.4.3            grid_4.4.2            urlchecker_1.0.1      profvis_0.4.0        
[33] xtable_1.8-4          colorspace_2.1-1      ggplot2_3.5.1         scales_1.3.0         
[37] iterators_1.0.14      MASS_7.3-64           cli_3.6.3             generics_0.1.3       
[41] remotes_2.5.0         rstudioapi_0.17.1     httr_1.4.7            tzdb_0.4.0           
[45] sessioninfo_1.2.2     cachem_1.1.0          profmem_0.6.0         stringr_1.5.1        
[49] splines_4.4.2         parallel_4.4.2        base64enc_0.1-3       vctrs_0.6.5          
[53] devtools_2.4.5        Matrix_1.7-1          benchmarkme_1.0.8     gratia_0.10.0        
[57] hms_1.1.3             patchwork_1.3.0       foreach_1.5.2         tidyr_1.3.1          
[61] benchmarkmeData_1.0.4 glue_1.8.0            codetools_0.2-20      ggokabeito_0.1.0     
[65] mvnfast_0.2.8         stringi_1.8.4         gtable_0.3.6          later_1.4.1          
[69] aws.s3_0.3.21         itsadug_2.4.1         munsell_0.5.1         tibble_3.2.1         
[73] pillar_1.10.1         htmltools_0.5.8.1     brio_1.1.5            R6_2.5.1             
[77] doParallel_1.0.17     rprojroot_2.0.4       shiny_1.10.0          lattice_0.22-6       
[81] readr_2.1.5           memoise_2.0.1         httpuv_1.6.15         Rcpp_1.0.14          
[85] nlme_3.1-166          mgcv_1.9-1            fs_1.6.5              forcats_1.0.0        
[89] usethis_3.1.0         plotfunctions_1.4     pkgconfig_2.0.3      
```