@arokem 2025-01-30
Thanks for taking a look! Heads up that this module is not currently on the main path of the most-used public APIs. It's here because we used it to sanity check some of the results early on in the development of the library, and I can see it having some uses down the line (e.g., to check assumptions), so there are potentially reasons to keep it working well.